Update Home.qml
v0.3.1
Removed 'Reauthorized' notifiers as they're unnecessary.
Removed the 301 response possibility. This was used when I had mjeb.dev redirect to an EvenNode server and wasn't sure if there'd be a change in service down the line. Now have domain and NodeJS app with DigitalOcean so there'll not be any need for a redirect.